Added a RomanNumeral interface and corresponding array of values
Defined a simple `RomanNumeral` interface that pairs a numeric value
with the Roman symbol. Also defined a constant array of all the values.
Exported the interface and array from a `lib` module.

Refactored the `integerToRoman` module to derive its own constants from
the shared array.

Added a TODO comment in the `index` module to restructure the modules.